Move test compressor to its own file
